refactor(teamcity): refactor TeamCity configs, update Qodana version
