--- .teamcity/patches/buildTypes/QodanaAnalysis.kts
@@ -1,31 +1,0 @@
-package patches.buildTypes
-
-import jetbrains.buildServer.configs.kotlin.v2018_2.*
-import jetbrains.buildServer.configs.kotlin.v2018_2.buildSteps.Qodana
-import jetbrains.buildServer.configs.kotlin.v2018_2.buildSteps.qodana
-import jetbrains.buildServer.configs.kotlin.v2018_2.ui.*
-
-/*
-This patch script was generated by TeamCity on settings change in UI.
-To apply the patch, change the buildType with id = 'QodanaAnalysis'
-accordingly, and delete the patch script.
-*/
-changeBuildType(RelativeId("QodanaAnalysis")) {
-    expectSteps {
-        qodana {
-            name = "Run Qodana for JS"
-            linter = customLinter {
-                image = "jetbrains/qodana-js:2023.2"
-            }
-            additionalDockerArguments = "-e QODANA_TOKEN=%env.QODANA_TOKEN%"
-            collectAnonymousStatistics = true
-        }
-    }
-    steps {
-        update<Qodana>(0) {
-            clearConditions()
-            inspectionProfile = default()
-            param("report-as-test", "")
-        }
-    }
-}
