--- .teamcity/patches/buildTypes/UnitTestsAndBuild.kts
@@ -1,81 +1,0 @@
-package patches.buildTypes
-
-import jetbrains.buildServer.configs.kotlin.v2018_2.*
-import jetbrains.buildServer.configs.kotlin.v2018_2.buildFeatures.commitStatusPublisher
-import jetbrains.buildServer.configs.kotlin.v2018_2.buildSteps.ScriptBuildStep
-import jetbrains.buildServer.configs.kotlin.v2018_2.buildSteps.script
-import jetbrains.buildServer.configs.kotlin.v2018_2.ui.*
-
-/*
-This patch script was generated by TeamCity on settings change in UI.
-To apply the patch, change the buildType with id = 'UnitTestsAndBuild'
-accordingly, and delete the patch script.
-*/
-changeBuildType(RelativeId("UnitTestsAndBuild")) {
-    expectSteps {
-        script {
-            name = "Test And Build"
-            scriptContent = """
-                #!/bin/bash
-                set -e -x
-                
-                node -v
-                npm -v
-                chown -R root:root . # See https://github.com/npm/cli/issues/4589
-                
-                mkdir node_modules
-                npm install
-                
-                npm run typecheck-ci
-                npm run test-ci
-                npm run build
-                npm run build-stories
-            """.trimIndent()
-            dockerImage = "node:22"
-            dockerImagePlatform = ScriptBuildStep.ImagePlatform.Linux
-            dockerRunParameters = "-v %teamcity.build.workingDir%/npmlogs:/root/.npm/_logs"
-        }
-    }
-    steps {
-        update<ScriptBuildStep>(0) {
-            clearConditions()
-            scriptContent = """
-                #!/bin/bash
-                set -e -x
-                
-                node -v
-                npm -v
-                chown -R root:root . # See https://github.com/npm/cli/issues/4589
-                
-                mkdir node_modules
-                npm install
-                
-                npm run typecheck-ci
-                npm run test-ci || true
-                npm run build
-                npm run build-stories
-            """.trimIndent()
-            param("org.jfrog.artifactory.selectedDeployableServer.downloadSpecSource", "Job configuration")
-            param("org.jfrog.artifactory.selectedDeployableServer.useSpecs", "false")
-            param("org.jfrog.artifactory.selectedDeployableServer.uploadSpecSource", "Job configuration")
-            param("teamcity.kubernetes.executor.pull.policy", "")
-        }
-    }
-
-    features {
-        remove {
-            commitStatusPublisher {
-                enabled = false
-                publisher = upsource {
-                    serverUrl = "https://upsource.jetbrains.com"
-                    projectId = "ring-ui"
-                    userName = "TeamCityReporter"
-                    password = "credentialsJSON:9eaa3cf0-4b14-49db-83f2-b141b3721922"
-                }
-            }
-        }
-    }
-
-    expectDisabledSettings("BUILD_EXT_5")
-    updateDisabledSettings("BUILD_EXT_4")
-}
